+++ src/abstract-property-plugin.js
@@ -35,7 +35,7 @@ export default class AbstractPropertyPlugin {
    * @returns {*}
    */
   getProperty(property, fallback) {
-    let segments = property.split('.');
+    let segments = property.toString().split('.');
     let scope = this._properties;
 
     for (let segment of segments) {
@@ -44,7 +44,6 @@ export default class AbstractPropertyPlugin {
         return this._handleNotFoundProperty(property, fallback);
       }
     }
-
     return this._propertyResolver.resolve(scope);
   }
 
@@ -55,7 +54,7 @@ export default class AbstractPropertyPlugin {
    * @private
    */
   _handleNotFoundProperty(property, fallback) {
-    if (typeof fallback === 'undefined') {
+    if (typeof fallback === 'undefined' || fallback === null) {
       throw new Error(`Property ${property} not found.`);
     }