+++ src/bsi-sass-property-plugin.js
@@ -1,22 +1,22 @@
 import AbstractPropertyPlugin from './abstract-property-plugin';
 
 export default class BsiSassPropertyPlugin extends AbstractPropertyPlugin {
+ 
   /**
-   * @param {*} property
-   * @param {*} fallback
-   * @returns {*}
+   * Sass function "bsiProperty()"
+   * 
+   * @param {[string, any]} param0 
+   * @returns 
    */
-  getProperty(property, fallback) {
-    let propertyName = property.getValue();
-
-    let value = super.getProperty(propertyName, fallback);
-
+  getSassProperty([property, fallback = null]) {
+    property = property.toString().replaceAll('"', '')
+    let value = super.getProperty(property, fallback);
     return typeof value.getSassObject === 'function' ? value.getSassObject() : value;
   }
 
   getFunction() {
     return {
-      'bsiProperty($property, $fallback: null)': this.getProperty.bind(this)
+      'bsiProperty($property, $fallback: null)': this.getSassProperty.bind(this)
     }
   }
 }