+++ src/webpack-config-builder.js
@@ -18,6 +18,7 @@ import BsiCxTwigContextWebpackPlugin from './bsi-cx-twig-context-webpack-plugin'
 import BsiLessPropertyPlugin from './bsi-less-property-plugin';
 import BuildContext from './build-context';
 import BsiSassPropertyPlugin from './bsi-sass-property-plugin';
+import PropertiesToScssConverter from './bsi-sass-properties-to-scss';
 import QueryConstant from './query-constant';
 import DistFolder from './dist-folder';
 import * as Version from './version';
@@ -304,10 +305,9 @@ export default class WebpackConfigBuilder {
             options: {
               sourceMap: true,
               sassOptions: {
-                functions: {
-                  ...new BsiSassPropertyPlugin(this.context).getFunction()
-                }
-              }
+                functions: new BsiSassPropertyPlugin(this.context).getFunction()
+              },
+              additionalData: new PropertiesToScssConverter(this.context).scssData
             }
           }
         ]