+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## latest
 * Add studioLink and studioLinkEnabled
+* Update sass-loader to get rid of warnings
+* Make all properties available as scss variables during the build
 
 
 ## [2.0.0] - 09.10.2025